requirements.txt
twikit==2.1.3
pandas==2.2.2
openai==1.43.0
python-dotenv==1.0.0
